[update] バージョン確認用APIの追加
