+++ backend/internal/model/schema.go
@@ -85,3 +85,8 @@ type Game struct {
 	Name    string `json:"name" db:"name"`
 	Process string `json:"process" db:"process"`
 }
+
+type UpdateResponse struct {
+	Current string `json:"current" db:"current"`
+	Support []string `json:"support" db:"support"`
+}
